+++ examples/lib/stories/parallax/advanced.dart
@@ -22,7 +22,6 @@ class AdvancedParallaxGame extends BaseGame {
     final parallax = ParallaxComponent.fromParallax(
       Parallax(
         await Future.wait(layers),
-        size,
         baseVelocity: Vector2(20, 0),
       ),
     );