+++ examples/lib/stories/parallax/small_parallax.dart
@@ -1,19 +1,10 @@
 import 'package:flame/components.dart';
 import 'package:flame/game.dart';
-import 'package:flame/parallax.dart';
 
-class ComponentParallaxGame extends BaseGame {
+class SmallParallaxGame extends BaseGame {
   @override
   Future<void> onLoad() async {
-    add(MyParallaxComponent());
-  }
-}
-
-class MyParallaxComponent extends ParallaxComponent
-    with HasGameRef<ComponentParallaxGame> {
-  @override
-  Future<void> onLoad() async {
-    parallax = await gameRef.loadParallax(
+    final component = await loadParallaxComponent(
       [
         'parallax/bg.png',
         'parallax/mountain-far.png',
@@ -21,8 +12,11 @@ class MyParallaxComponent extends ParallaxComponent
         'parallax/trees.png',
         'parallax/foreground-trees.png',
       ],
+      size: Vector2.all(200),
       baseVelocity: Vector2(20, 0),
       velocityMultiplierDelta: Vector2(1.8, 1.0),
-    );
+    )
+      ..position = size / 2;
+    add(component);
   }
 }